+++ sdk/nodejs/runtime/resource.ts
@@ -49,7 +49,6 @@ import {
 import {
     excessiveDebugOutput,
     getMonitor,
-    getProject,
     getRootResource,
     getStack,
     isDryRun,
@@ -60,7 +59,6 @@ import {
 } from "./settings";
 
 const gstruct = require("google-protobuf/google/protobuf/struct_pb.js");
-const providerproto = require("../proto/provider_pb.js");
 const resproto = require("../proto/resource_pb.js");
 
 interface ResourceResolverOperation {
@@ -568,13 +566,14 @@ async function prepareResource(label: string, res: Resource, custom: boolean, re
         // The list of all dependencies (implicit or explicit).
         const allDirectDependencies = new Set<Resource>(explicitDirectDependencies);
 
-        const allDirectDependencyURNs = await getAllTransitivelyReferencedResourceURNs(explicitDirectDependencies);
+        const exclude = new Set<Resource>([res]);
+        const allDirectDependencyURNs = await getAllTransitivelyReferencedResourceURNs(explicitDirectDependencies, exclude);
         const propertyToDirectDependencyURNs = new Map<string, Set<URN>>();
 
         for (const [propertyName, directDependencies] of propertyToDirectDependencies) {
             addAll(allDirectDependencies, directDependencies);
 
-            const urns = await getAllTransitivelyReferencedResourceURNs(directDependencies);
+            const urns = await getAllTransitivelyReferencedResourceURNs(directDependencies, exclude);
             addAll(allDirectDependencyURNs, urns);
             propertyToDirectDependencyURNs.set(propertyName, urns);
         }
@@ -619,7 +618,7 @@ function addAll<T>(to: Set<T>, from: Set<T>) {
 }
 
 /** @internal */
-export async function getAllTransitivelyReferencedResourceURNs(resources: Set<Resource>): Promise<Set<string>> {
+export async function getAllTransitivelyReferencedResourceURNs(resources: Set<Resource>, exclude: Set<Resource>): Promise<Set<string>> {
     // Go through 'resources', but transitively walk through **Component** resources, collecting any
     // of their child resources.  This way, a Component acts as an aggregation really of all the
     // reachable resources it parents.  This walking will stop when it hits custom resources.
@@ -653,7 +652,7 @@ export async function getAllTransitivelyReferencedResourceURNs(resources: Set<Re
     // Then we filter to only include Custom and Remote resources.
     const transitivelyReachableCustomResources =
         [...transitivelyReachableResources]
-            .filter(r => CustomResource.isInstance(r) || (r as ComponentResource).__remote);
+            .filter(r => (CustomResource.isInstance(r) || (r as ComponentResource).__remote) && !exclude.has(r));
     const promises = transitivelyReachableCustomResources.map(r => r.urn.promise());
     const urns = await Promise.all(promises);
     return new Set<string>(urns);