+++ sdk/nodejs/tests/runtime/langhost/run.spec.ts
@@ -14,7 +14,6 @@
 
 import * as assert from "assert";
 import * as childProcess from "child_process";
-import * as os from "os";
 import * as path from "path";
 import { ID, runtime, URN } from "../../../index";
 import { asyncTest } from "../../util";