+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-## 3.33.2 (2022-05-26)
+## 3.33.2 (2022-05-27)
 
 ### Improvements
 
@@ -18,6 +18,9 @@ CHANGELOG
 
 - [sdk/python] Fix spurious diffs causing an "update" on resources created by dynamic providers.
   [#9656](https://github.com/pulumi/pulumi/pull/9656)
+
+- [sdk/python] Do not depend on the children of remote component resources.
+  [#9665](https://github.com/pulumi/pulumi/pull/9665)
 
 - [codegen/nodejs] Emit the "package.json".pulumi.server as "server" correctly. Previously, "pluginDownloadURL" was emitted but never read.
   [#9662](https://github.com/pulumi/pulumi/pull/9662)
@@ -36,6 +39,9 @@ CHANGELOG
 
 - [sdk/python] Pin protobuf library to <4.
   [#9696](https://github.com/pulumi/pulumi/pull/9696)
+
+- [sdk/proto] Inline dockerfile used to generate protobuf code.
+  [#9700](https://github.com/pulumi/pulumi/pull/9700)
 
 ## 3.33.1 (2022-05-19)